pyproject.toml
[project]
name = "baserow-aksioma-payment-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.27.0",
    "aiohttp>=3.13.5",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "rapidfuzz>=3.14.5",
    "tenacity>=9.1.4",
]
